Add some collision detection methods into the frogger.js file.
Update the frogger.js with a new collisionDetection method that passed in the full moverArray and runs a forEach to detect if the square of the frog is interacting with any of the squares pulled from the movers array. If the type of moverArray is Car then set the color of frog to red.