Add functionality for start and reset buttons.
Create button in html / css for the start and reset buttons that sit above the canvas. Add functionality in the index.js file to run the start and reset buttons. The start method will contain the gameloop, allowing the game to run on start button click and not page load.